+++ frontend/src/pages/planner/Planner.tsx
@@ -2,9 +2,9 @@ import { Spinner } from '../../components/Spinner'
 import ErrorTray from './ErrorTray'
 import PlanBoard from './planBoard/PlanBoard'
 import ControlTopBar from './ControlTopBar'
-import CourseSelectorDialog from './CourseSelectorDialog'
-import LegendModal from './LegendModal'
-import SavePlanModal from './SavePlanModal'
+import CourseSelectorDialog from './dialogs/CourseSelectorDialog'
+import LegendModal from './dialogs/LegendModal'
+import SavePlanModal from './dialogs/SavePlanModal'
 import CurriculumSelector from './CurriculumSelector'
 import AlertModal from '../../components/AlertModal'
 import { useParams, Navigate } from '@tanstack/react-router'
@@ -21,7 +21,7 @@ import { HTML5Backend } from 'react-dnd-html5-backend'
 import { collectRequirements, handleErrors, PlannerStatus } from './utils/utils'
 import { updateCurriculum, isMinorValid, isMajorValid, loadCurriculumsData } from './utils/CurriculumUtils'
 import ReceivePaste from './utils/ReceivePaste'
-import ModBanner from '../mod/ModBanner'
+import ModBanner from './ModBanner'
 
 const reduceCourseDetails = (old: Record<string, PseudoCourseDetail>, add: Record<string, PseudoCourseDetail>): Record<string, PseudoCourseDetail> => {
   return { ...old, ...add }
