+++ frontend/src/pages/planner/dialogs/LegendModal.tsx
